chore: move all execAndWaitResponse related functions to single file
